Update firerecord
Added Key Listener for specific Keys. Edits to HandleType, HandleEvent,
run(recording click and type events) and fire. Altered the format to
handle any properties associated with the element. Fire also confirms
every property. Changed the recording process to also include if it was
a click, typing, or key event to result in easier playback. removed a
few functions also.